move EntityTable class to right place
