+++ cdap-watchdog/src/main/java/co/cask/cdap/metrics/guice/MetricsHandlerModule.java
@@ -16,7 +16,6 @@
 
 package co.cask.cdap.metrics.guice;
 
-import co.cask.cdap.api.metrics.MetricStore;
 import co.cask.cdap.common.conf.Constants;
 import co.cask.cdap.gateway.handlers.CommonHandlers;
 import co.cask.cdap.logging.gateway.handlers.LogHandler;
@@ -27,9 +26,6 @@ import co.cask.cdap.metrics.query.MetricsDiscoveryHandler;
 import co.cask.cdap.metrics.query.MetricsHandler;
 import co.cask.cdap.metrics.query.MetricsQueryHandler;
 import co.cask.cdap.metrics.query.MetricsQueryService;
-import co.cask.cdap.metrics.store.DefaultMetricDatasetFactory;
-import co.cask.cdap.metrics.store.DefaultMetricStore;
-import co.cask.cdap.metrics.store.MetricDatasetFactory;
 import co.cask.http.HttpHandler;
 import com.google.inject.PrivateModule;
 import com.google.inject.Scopes;
