+++ cdap-watchdog/src/main/java/co/cask/cdap/metrics/store/cube/CubeQuery.java
@@ -21,7 +21,6 @@ import co.cask.cdap.metrics.store.timeseries.MeasureType;
 import com.google.common.base.Joiner;
 import com.google.common.base.Objects;
 import com.google.common.collect.ImmutableList;
-import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Maps;
 
 import java.util.List;
