+++ cdap-watchdog/src/main/java/co/cask/cdap/metrics/store/timeseries/FactTable.java
@@ -22,7 +22,6 @@ import co.cask.cdap.api.metrics.TagValue;
 import co.cask.cdap.common.utils.ImmutablePair;
 import co.cask.cdap.data2.dataset2.lib.table.FuzzyRowFilter;
 import co.cask.cdap.data2.dataset2.lib.table.MetricsTable;
-import co.cask.cdap.metrics.data.EntityTable;
 import com.google.common.base.Function;
 import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
