+++ cdap-watchdog/src/test/java/co/cask/cdap/metrics/store/timeseries/EntityTableTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package co.cask.cdap.metrics.data;
+package co.cask.cdap.metrics.store.timeseries;
 
 import co.cask.cdap.data2.dataset2.lib.table.MetricsTable;
 import co.cask.cdap.data2.dataset2.lib.table.inmemory.InMemoryMetricsTable;
